requirements.txt
flask
flask-cors
python-dotenv
cohere
torch
whisper
